+++ src/main/java/com/example/expensetracker/entity/Income.java
@@ -10,7 +10,7 @@ public class Income {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String source;
     private BigDecimal amount;
     @ManyToOne(fetch = FetchType.EAGER)
@@ -20,13 +20,7 @@ public class Income {
     public Income() {
     }
 
-    public Income(
-            Integer id,
-            String source,
-            BigDecimal amount,
-            Tag tag,
-            LocalDate date
-    ) {
+    public Income(Long id, String source, BigDecimal amount, Tag tag, LocalDate date) {
         this.id = id;
         this.source = source;
         this.amount = amount;
@@ -34,11 +28,11 @@ public class Income {
         this.date = date;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
