+++ src/main/java/com/example/expensetracker/entity/Tag.java
@@ -6,21 +6,21 @@ import jakarta.persistence.*;
 public class Tag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String title;
     public Tag() {
     }
 
-    public Tag(Integer id, String title) {
+    public Tag(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
